requirements.txt
line-bot-sdk
flask
configparser
openai
pandas